requirements.txt
# Core data processing
pandas>=2.2.0
numpy>=1.26.0
python-dateutil>=2.8.0

# Graph analysis
networkx>=3.2.0
python-louvain>=0.16

# NLP and similarity
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Interactive dashboard
streamlit>=1.28.0

# Utilities
tqdm>=4.66.0
pydantic>=2.5.0
typing-extensions>=4.9.0

# Optional: Interactive visualization
pyvis>=0.3.0
